Register jinja filters in app factory
